Enhance student validation and database schema
Updated `AddStudentValidator` and `EditStudentValidator` to include validation for new `NameEn` and `NameAr` properties. Modified `BaseStudentCommand` to support these properties, along with corresponding validation rules. Adjusted `StudentMapper` and `EditStudentMapping` to handle new name fields. Changed `Grade` type in `StudentsSubjects` from `int?` to `decimal?` for precision. Updated `InstructorConfiguration` for salary precision and revised migration files to reflect schema changes, including new configurations in `StudentsSubjectsConfiguration`.

+++ SchoolProject.Domain/Entities/StudentsSubjects.cs
@@ -7,7 +7,7 @@ public class StudentsSubjects
 
     public int SubID { get; set; }
 
-    public int? Grade { get; set; }
+    public decimal? Grade { get; set; }
 
     public virtual Student Student { get; set; } = null!;
